Update convertHandler.js
Modified the logic in the .getNum(input) method using multiple checks based on the unit of measurement index and catching syntax errors or new Error() instances when the result is evaluated using math.js's math.eval() method. Added the math.js library as a dependency to avoid using the unsecured javaScript vanilla eval() function. Added extra functionality to the project, by introducing 6 extra units of measurement in the main constructor logic. The project is now final and passed both completed suite of tests (1_unit-tests and 2_functional-tests). Ready to be submitted to FCC for review if necessary.